feat: move `get-image.aspx` to the api (api/resources/image)
