+++ aspx/wwwroot/App_Code/api/resources/Controller.cs
@@ -1,0 +1,9 @@
+using System.Web.Http;
+
+namespace RAWebServer.Api
+{
+  [RoutePrefix("api/resources")]
+  public partial class ResourceController : ApiController
+  {
+  }
+}
