--- aspx/wwwroot/get-image.aspx
@@ -1,1 +1,0 @@
-<%@ Page Language="C#" AutoEventWireup="true" CodeFile="get-image.aspx.cs" Inherits="GetImage" %>
