+++ frontend/app.html
@@ -19,7 +19,7 @@
   <div class="root-splash-wrapper">
     <div class="root-titlebar">
       <div class="left">
-        <img src="%raweb.base%lib/assets/icon.svg" alt="" class="logo" />
+        <img src="%raweb.base%/lib/assets/icon.svg" alt="" class="logo" />
         <span class="title">RAWeb</span>
       </div>
       <div class="right"></div>
